bands aren't floats
git-svn-id: https://svn.osgeo.org/gdal/trunk@16912 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/apps/gdaldem.cpp
@@ -756,7 +756,7 @@ int main( int argc, char ** argv )
              EQUAL(argv[i], "-b"))
           )
         {
-            nBand = atof(argv[++i]);
+            nBand = atoi(argv[++i]);
         }
         else if ( EQUAL(argv[i], "-quiet") )
         {
